solar_prescription/solar_prescription/requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
Werkzeug==3.1.3
Jinja2==3.1.4
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.7
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3
pandas==2.2.2

# UI smoke tests
playwright==1.50.0
